requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=6.5.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
mlflow>=2.8.0  # for experiment tracking
nltk==3.9.1
tabulate==0.9.0
pyyaml
torchvision==0.21.0
